// packages/ui/src/components/ui/form_disabled/form.types.ts
export type FieldValues = Record<string, any>
// ...
export interface BaseFormFieldProps {
  name: string
  label?: JSX.Element
  description?: JSX.Element
  required?: boolean
  disabled?: boolean
  placeholder?: string
  autocomplete?: string
}

/**
 * Form field component props
 */
export interface FormFieldProps<T extends FieldValues = FieldValues>
  extends BaseFormFieldProps {
  class?: string
  layout?: 'vertical' | 'horizontal'
  validationState?: ValidationState
  error?: string
  helpText?: string
  tooltip?: JSX.Element
  icon?: JSX.Element
  children: JSX.Element | ((field: any) => JSX.Element)
}